Allows to override user-agent header in a cred file
See also: #70
