+++ changeme/core.py
@@ -4,7 +4,6 @@ from changeme.redis_queue import RedisQueue
 import logging
 from logutils import colorize
 import os
-import random
 import re
 import redis
 from .report import Report
@@ -182,7 +181,7 @@ class Config(object):
         if self.debug or self.validate:
             logger.setLevel(logging.DEBUG)
 
-        self.useragent = {'User-Agent': self.useragent if self.useragent else get_useragent()}
+        self.useragent = {'User-Agent': self.useragent} if self.useragent else {}
 
         if ',' in self.protocols:
             self.protocols = self.protocols.split(',')
@@ -371,23 +370,6 @@ def print_creds(creds):
         print("\n%s (%s)" % (cred['name'], cred['category']))
         for i in cred['auth']['credentials']:
             print("  - %s:%s" % (i['username'], i['password']))
-
-
-def get_useragent():
-    headers_useragents = [
-        'Mozilla/5.0 (X11; U; Linux x86_64; en-US; rv:1.9.1.3) Gecko/20090913 Firefox/3.5.3',
-        'Mozilla/5.0 (Windows; U; Windows NT 6.1; en; rv:1.9.1.3) Gecko/20090824 Firefox/3.5.3 (.NET CLR 3.5.30729)',
-        'Mozilla/5.0 (Windows; U; Windows NT 5.2; en-US; rv:1.9.1.3) Gecko/20090824 Firefox/3.5.3 (.NET CLR 3.5.30729)',
-        'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.1.1) Gecko/20090718 Firefox/3.5.1',
-        'Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/532.1 (KHTML, like Gecko) Chrome/4.0.219.6 Safari/532.1',
-        'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; WOW64; Trident/4.0; SLCC2; .NET CLR 2.0.50727; InfoPath.2)',
-        'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 5.2; Win64; x64; Trident/4.0)',
-        'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 5.1; Trident/4.0; SV1; .NET CLR 2.0.50727; InfoPath.2)',
-        'Mozilla/5.0 (Windows; U; MSIE 7.0; Windows NT 6.0; en-US)',
-        'Mozilla/4.0 (compatible; MSIE 6.1; Windows XP)',
-        'Opera/9.80 (Windows NT 5.2; U; ru) Presto/2.5.22 Version/10.51'
-    ]
-    return random.choice(headers_useragents)
 
 
 def check_for_interrupted_scan(config):
