+++ changeme/scanners/http_get.py
@@ -1,4 +1,5 @@
 import base64
+import random
 from requests import session
 from .scanner import Scanner
 import re
@@ -11,6 +12,24 @@ except ImportError:
     # Python 2
     from urllib import urlencode
     from urlparse import urlparse
+
+HEADERS_USERAGENTS = [
+    'Mozilla/5.0 (X11; U; Linux x86_64; en-US; rv:1.9.1.3) Gecko/20090913 Firefox/3.5.3',
+    'Mozilla/5.0 (Windows; U; Windows NT 6.1; en; rv:1.9.1.3) Gecko/20090824 Firefox/3.5.3 (.NET CLR 3.5.30729)',
+    'Mozilla/5.0 (Windows; U; Windows NT 5.2; en-US; rv:1.9.1.3) Gecko/20090824 Firefox/3.5.3 (.NET CLR 3.5.30729)',
+    'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.1.1) Gecko/20090718 Firefox/3.5.1',
+    'Mozilla/5.0 (Windows; U; Windows NT 5.1; en-US) AppleWebKit/532.1 (KHTML, like Gecko) Chrome/4.0.219.6 Safari/532.1',
+    'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; WOW64; Trident/4.0; SLCC2; .NET CLR 2.0.50727; InfoPath.2)',
+    'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 5.2; Win64; x64; Trident/4.0)',
+    'Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 5.1; Trident/4.0; SV1; .NET CLR 2.0.50727; InfoPath.2)',
+    'Mozilla/5.0 (Windows; U; MSIE 7.0; Windows NT 6.0; en-US)',
+    'Mozilla/4.0 (compatible; MSIE 6.1; Windows XP)',
+    'Opera/9.80 (Windows NT 5.2; U; ru) Presto/2.5.22 Version/10.51'
+]
+
+
+def get_useragent():
+    return random.choice(HEADERS_USERAGENTS)
 
 
 class HTTPGetScanner(Scanner):
@@ -25,10 +44,19 @@ class HTTPGetScanner(Scanner):
         self.response = None
 
         headers = self.cred['auth'].get('headers', dict())
+        custom_ua = False
         if headers:
             for h in headers:
                 self.headers.update(h)
-        self.headers.update(self.config.useragent)
+                if not custom_ua and any(k.lower() == 'user-agent' for k in h):
+                    custom_ua = True
+
+        # If set, take user agent from CLI args, otherwise, pick a random
+        # one if not provided in the cred file.
+        if self.config.useragent:
+            self.headers.update(self.config.useragent)
+        elif not custom_ua:
+            self.headers.update({'User-Agent': get_useragent()})
 
         # make the cred have only one u:p combo
         self.cred['auth']['credentials'] = [{'username': self.username, 'password': self.password}]
